improve source map granularity
fixes #3023
